Add edit product link(Dev only)
